+++ src/main/java/com/powsybl/openloadflow/ac/TransformerVoltageControlOuterLoop.java
@@ -92,7 +92,7 @@ public class TransformerVoltageControlOuterLoop extends AbstractTransformerVolta
                 branch.setVoltageControlEnabled(true);
                 status = OuterLoopStatus.UNSTABLE;
             }
-            checkControl(context.getNetwork());
+            context.getNetwork().fixTransformerVoltageControls();
         }
 
         // At second outer loop iteration, the transformers are rounded. The generator voltage controls that were
